+++ NEWS.md
@@ -4,6 +4,10 @@
 
 * Removed function 'var_rename()', which is in pkg 'sjmisc'.
 
+## New functions
+
+* `get_term_labels()` and `get_dv_labels()` to retrieve term labels from regression models.
+
 ## Changes to functions
 
 * `as_numeric()` gets a `use.labels`-argument to use value labels as new values if these are numeric.